leopold/fc980c: Move existing config to tmk subtarget
