+++ keyboards/leopold/fc980c/tmk/info.json
@@ -1,6 +1,5 @@
 {
-  "keyboard_name": "FC980C",
-  "manufacturer": "Hasu",
+  "keyboard_name": "FC980C with tmk controller",
   "url": "",
   "maintainer": "qmk",
   "usb": {
