+++ keyboards/leopold/fc980c/tmk/tmk.c
@@ -15,7 +15,7 @@ You should have received a copy of the GNU General Public License
 along with this program.  If not, see <http://www.gnu.org/licenses/>.
 */
 
-#include "fc980c.h"
+#include "tmk.h"
 
 #ifdef ACTUATION_DEPTH_ADJUSTMENT
 void matrix_init_kb(void) {
