+++ keyboards/leopold/info.json
@@ -1,0 +1,3 @@
+{
+  "manufacturer": "Leopold"
+}
